[yarn]: Enable support for corepack offline installs
This is required to enable building from nix, as corepack cannot fetch the yarn version from npmjs.com during a build.
Thus, an offline version is required.

This also means that when upgrading yarn, also ensure to run `yarn fetch-yarn` to fetch the latest yarn version to the correct place, and commit that with your version bump.
This will ensure that the version in the repo is kept in sync with the version in package.json.